- Changed SmsToMobile.MsgChargeUSD from double to float - Optimized RepeatCharges.GetAll()
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@12518 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Misc/ConvertDatabases3.cs
@@ -8445,7 +8445,7 @@ namespace OpenDentBusiness {
 						MsgText text NOT NULL,
 						SmsStatus tinyint NOT NULL,
 						MsgParts int NOT NULL,
-						MsgChargeUSD double NOT NULL,
+						MsgChargeUSD float NOT NULL,
 						ClinicNum bigint NOT NULL,
 						CustErrorText varchar(255) NOT NULL,
 						DateTimeSent datetime NOT NULL DEFAULT '0001-01-01 00:00:00',
@@ -8943,7 +8943,6 @@ namespace OpenDentBusiness {
 
 	}
 }
-
